orchestrator/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
grpcio==1.70.0
grpcio-tools==1.70.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
protobuf==5.29.3
Werkzeug==3.0.1
Flask-CORS==4.0.0
watchdog==6.0.0
marshmallow==3.26.1
requests==2.31.0
pgvector==0.3.6
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
python-dotenv==1.0.1
